Move richDEM wrapper functions, move terrain ruggedness to maintain consistent order of roughness indexes, and fix minor docstrings issues
